Add a folder to organize testcase files
